requirements.txt
numpy 
pandas
scipy
matplotlib
torch
torchvision
torchaudio
umap-learn
scikit-learn